refactor(tmux): cleaner capture post-processing
